chore: car 패키지 생성
